Reorganize template files
