+++ openapi/templates/Client.cs.hbs
