+++ openapi/templates/Enum.cs.hbs
