+++ openapi/templates/IClient.cs.hbs
