+++ openapi/templates/IModel.cs.hbs
