+++ openapi/templates/Model.cs.hbs
