+++ openapi/templates/Resolver.cs.hbs
